*Added override specifier. *Corrected precompiled header. *Fixed compile warning from converting size_t to uint32_t.

+++ GraphAlgorithms/DirectedGraph.cpp
@@ -44,7 +44,7 @@ uint32_t DirectedGraph::getOutDegree(uint32_t vertex) const
 	if (!isValidVertex(vertex))
 		return 0;
 
-	return _adjacencyList[vertex].size();
+	return (uint32_t)_adjacencyList[vertex].size();
 }
 
 uint32_t DirectedGraph::getMinDegree() const
